chore(@clayui/table): update Table component stories
